pyproject.toml
[tool.poetry]
name = "nightskycam"
version = "0.1.0"
description = "taking pictures at night"
authors = ["Vincent Berenz <vberenz@tuebingen.mpg.de>"]
packages = [{ include = "nightskycam" }]

[tool.poetry.scripts]
nightskycam-start = 'nightskycam.main:execute'
nightskycam-test-status-server = 'nightskycam.status.test_server:run_status_test_server'
nightskycam-repetitive-test = 'nightskycam.main:repetitive_starting_test'

[tool.poetry.dependencies]
python = "^3.9"
astropy = "^6.0.1"
auto-stretch = "^0.0.1"
netifaces = "^0.11.0"
psutil = "^5.9.8"
pyephem = "^9.99"
pyftpdlib = "^1.5.9"
pyusb = "^1.2.1"
nptyping = "^2.5.0"
requests = "^2.31.0"
cachetools = "^5.3.3"
websocket-client = "^1.8.0"
matplotlib = "^3.8.4"
opencv-python = "^4.9.0.80"
websocket-server = "^0.6.4"
nightskyrunner = "^0.1.0"
nightskycam-serialization = "^0.1.0"
h5darkframes = "^0.1.4"
nightskycam-images = "^0.1.0"
typing-extensions = "^4.12.2"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pyflakes = "^3.2.0"
autopep8 = "^2.1.0"
mypy = "^1.10.0"
types-requests = "^2.31.0.20240406"
types-toml = "^0.10.8.20240310"
pytest = "^8.1.1"
pytest-mock = "^3.14.0"
types-cachetools = "^5.3.0.7"
types-psutil = "^5.9.5.20240423"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
